+++ CSharpGL.Windows/RenderContexts/HiddenWindowRenderContext.cs
@@ -16,7 +16,7 @@ namespace CSharpGL
         /// <param name="bitDepth">The bit depth.</param>
         /// <param name="parameter">The parameter</param>
         /// <returns></returns>
-        public HiddenWindowRenderContext(int width, int height, int bitDepth, object parameter)
+        public HiddenWindowRenderContext(int width, int height, short bitDepth, object parameter)
             : base(width, height, bitDepth, parameter)
         {
             // Create a new window class, as basic as possible.
